--- app/src/main/java/com/rubdev/myapplication/Model/GithubRepository.kt
@@ -1,10 +1,0 @@
-package com.rubdev.myapplication.Model
-
-data class GithubRepository(
-    val repositoryName: String,
-    val description: String,
-    val authorName: String,
-    val stars: String,
-    val forks: String
-
-)
