--- app/src/main/java/com/rubdev/myapplication/Model/pullRequestOfRepository.kt
@@ -1,10 +1,0 @@
-package com.rubdev.myapplication.Model
-
-import java.util.*
-
-data class pullRequestOfRepository(
-    val title: String,
-    val date: Date,
-    val body: String
-
-)
