--- app/src/main/java/com/rubdev/myapplication/Repository/GitHubApiService.kt
@@ -1,6 +1,0 @@
-package com.rubdev.myapplication.Repository
-
-interface GitHubApiService {
-
-
-}
