+++ app/src/main/res/drawable/ic_git_branch.xml
@@ -1,0 +1,10 @@
+<vector
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportHeight="24"
+    android:viewportWidth="24">
+    <path
+        android:fillColor="#FFDF930B"
+        android:pathData="M 21.081 7.505 C 21.081 5.842 19.746 4.508 18.081 4.508 C 16.416 4.508 15.081 5.842 15.081 7.505 C 15.081 8.599 15.696 9.573 16.581 10.082 L 16.581 10.532 C 16.551 11.311 16.236 12 15.636 12.599 C 15.036 13.199 14.346 13.513 13.566 13.543 C 12.321 13.573 11.346 13.783 10.566 14.218 L 10.566 7.085 C 11.451 6.576 12.066 5.617 12.066 4.508 C 12.066 2.845 10.731 1.512 9.066 1.512 C 7.401 1.512 6.081 2.845 6.081 4.508 C 6.081 5.602 6.696 6.576 7.581 7.085 L 7.581 16.915 C 6.696 17.439 6.081 18.398 6.081 19.492 C 6.081 21.155 7.416 22.488 9.081 22.488 C 10.746 22.488 12.081 21.155 12.081 19.492 C 12.081 18.698 11.781 17.993 11.286 17.454 C 11.421 17.364 12.006 16.84 12.171 16.75 C 12.546 16.585 13.011 16.495 13.581 16.495 C 15.156 16.42 16.506 15.821 17.706 14.622 C 18.906 13.423 19.506 11.655 19.581 10.097 L 19.551 10.097 C 20.466 9.558 21.081 8.599 21.081 7.505 Z M 9.081 2.71 C 10.071 2.71 10.881 3.534 10.881 4.508 C 10.881 5.482 10.056 6.306 9.081 6.306 C 8.106 6.306 7.281 5.482 7.281 4.508 C 7.281 3.534 8.106 2.71 9.081 2.71 Z M 9.081 21.305 C 8.091 21.305 7.281 20.481 7.281 19.507 C 7.281 18.533 8.106 17.709 9.081 17.709 C 10.056 17.709 10.881 18.533 10.881 19.507 C 10.881 20.481 10.056 21.305 9.081 21.305 Z M 18.081 9.318 C 17.091 9.318 16.281 8.494 16.281 7.52 C 16.281 6.546 17.106 5.722 18.081 5.722 C 19.056 5.722 19.881 6.546 19.881 7.52 C 19.881 8.494 19.056 9.318 18.081 9.318 Z"/>
+</vector>
