+++ app/src/main/res/drawable/shadow_ligth.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:angle="90"
+        android:endColor="#33333333"
+        android:startColor="@android:color/transparent"/>
+</shape>
