+++ app/src/main/res/layout/gradient_shadow_layout.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<View
+    android:id="@+id/gradient_shadow"
+    android:layout_width="match_parent"
+    android:layout_height="3dp"
+    android:background="@drawable/shadow_light"
+    app:layout_scrollFlags="scroll|enterAlways"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"/>
