+++ app/src/main/res/layout/repository_layout.xml
@@ -1,6 +1,122 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
     android:layout_height="match_parent">
+
+
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/name_repository"
+        android:textColor="#2979FF"
+        android:textSize="18sp"
+        app:layout_constraintEnd_toEndOf="@+id/descriptyon_repo"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/descriptyon_repo"
+        android:layout_width="250dp"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/description_repo"
+        android:textColor="@color/black"
+        android:textSize="12sp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/textView" />
+
+    <ImageView
+        android:id="@+id/image_star"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="16dp"
+        app:layout_constraintStart_toEndOf="@+id/number_stars"
+        app:layout_constraintTop_toBottomOf="@+id/descriptyon_repo"
+        app:srcCompat="@drawable/ic_baseline_star_24" />
+
+    <ImageView
+        android:id="@+id/image_fork"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="16dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/descriptyon_repo"
+        app:srcCompat="@drawable/ic_git_branch" />
+
+    <TextView
+        android:id="@+id/number_forks"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="16dp"
+        android:text="@string/number_fork"
+        android:textColor="#FFDF930B"
+        app:layout_constraintStart_toEndOf="@+id/image_star"
+        app:layout_constraintTop_toBottomOf="@+id/descriptyon_repo" />
+
+    <TextView
+        android:id="@+id/number_stars"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="16dp"
+
+        android:text="@string/number_stars"
+        android:textColor="#FFDF930B"
+        app:layout_constraintStart_toEndOf="@+id/image_fork"
+        app:layout_constraintTop_toBottomOf="@+id/descriptyon_repo" />
+
+    <ImageView
+        android:id="@+id/imageView5"
+        android:layout_width="wrap_content"
+        android:layout_height="70dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginEnd="8dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/descriptyon_repo"
+        app:layout_constraintTop_toTopOf="parent"
+        app:srcCompat="@drawable/ic_launcher_foreground" />
+
+    <TextView
+        android:id="@+id/username_repo"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="8dp"
+        android:text="@string/username"
+        android:textColor="#2979FF"
+        android:textSize="12dp"
+        app:layout_constraintEnd_toEndOf="@+id/imageView5"
+        app:layout_constraintStart_toStartOf="@+id/imageView5"
+        app:layout_constraintTop_toBottomOf="@+id/imageView5" />
+
+    <TextView
+        android:id="@+id/name_surname"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/name_surname"
+        android:textSize="12sp"
+        app:layout_constraintEnd_toEndOf="@+id/username_repo"
+        app:layout_constraintStart_toStartOf="@+id/username_repo"
+        app:layout_constraintTop_toBottomOf="@+id/username_repo" />
+
+    <include
+        layout="@layout/gradient_shadow_layout"
+        android:layout_width="match_parent"
+        android:layout_below="@id/name_surname"
+        android:layout_height="3dp"
+        android:layout_marginStart="08dp"
+        android:layout_marginLeft="08dp"
+        android:layout_marginTop="08dp"
+        android:layout_marginEnd="08dp"
+        android:layout_marginRight="08dp"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
